refactor: bridge 폴더 이전
