+++ web/src/lib/bridge/index.ts
@@ -3,10 +3,6 @@ import { linkBridge } from "@webview-bridge/web";
 import type { AppBridge } from "@/types/bridge/bridge.types";
 
 export const bridge = linkBridge<AppBridge>({
-  onReady: () => {
-    // 확인용
-    console.log("브릿지 연결 성공");
-  },
   initialBridge: {
     isLoggedIn: false,
   },
